Move initialize function for attitude
